Renaming of classes and interfaces
Stuff -> StuffImpl
IStuff -> Stuff
