+++ src/main/java/analysis/AnalysisProcessor.java
@@ -4,7 +4,7 @@ package main.java.analysis;
  * Processors should be classes which manipulate the analysis in some way. Plugins should be used
  * here.
  */
-public interface IAnalysisProcessor {
+public interface AnalysisProcessor {
 
 	/**
 	 * Manipulates the analysis in some way, independent of the analysis.
