--- src/main/java/distributed/IConnectionManagerORM.java
@@ -1,12 +1,0 @@
-package main.java.distributed;
-
-import java.sql.SQLException;
-
-import com.j256.ormlite.support.ConnectionSource;
-
-public interface IConnectionManagerORM extends IConnectionManager {
-
-
-	public ConnectionSource getConnectionORM() throws SQLException;
-
-}
