+++ src/main/java/distributed/results/ResultProcessorException.java
@@ -2,7 +2,7 @@ package main.java.distributed.results;
 
 @SuppressWarnings("serial")
 /**
- * The class for all the possible exceptions in the ResultProcessor class
+ * The class for all the possible exceptions in the ResultProcessorImpl class
  * In order to show the exceptions friendly to the user
  */
 public class ResultProcessorException extends RuntimeException {
