+++ src/test/java/distributed/workload/TestWorkloadDAO.java
@@ -12,13 +12,13 @@ import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
-import main.java.distributed.ConnectionManager;
+import main.java.distributed.ConnectionManagerImpl;
 import main.java.distributed.workload.WorkTask;
-import main.java.distributed.workload.WorkloadDAO;
+import main.java.distributed.workload.WorkloadDaoImpl;
 
 import org.junit.Test;
 
-public class TestWorkloadDAO {
+public class TestWorkloadDao {
 	
 	
 	
@@ -27,7 +27,7 @@ public class TestWorkloadDAO {
 		// Result set 
 		List<WorkTask> expected = new ArrayList<WorkTask>();
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		ResultSet results = mock(ResultSet.class);
@@ -38,7 +38,7 @@ public class TestWorkloadDAO {
 		when(statement.executeQuery(anyString())).thenReturn(results);
 		when(results.next()).thenReturn(false);
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		List<WorkTask> finalReturn = wldao.retrieveWork(expected.size());
 		assertEquals(finalReturn, expected);
 	}
@@ -50,7 +50,7 @@ public class TestWorkloadDAO {
 		WorkTask wt1 = new WorkTask(1, new URL("http://1.com"));
 		expected.add(wt1);
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		ResultSet results = mock(ResultSet.class);
@@ -63,7 +63,7 @@ public class TestWorkloadDAO {
 		when(results.getInt("id")).thenReturn(wt1.getId());
 		when(results.getString("url")).thenReturn(wt1.getURL().toString());
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		List<WorkTask> finalReturn = wldao.retrieveWork(expected.size());
 		assertEquals(finalReturn, expected);
 	}
@@ -76,7 +76,7 @@ public class TestWorkloadDAO {
 		WorkTask wt2 = new WorkTask(2, new URL("http://2.com"));
 		expected.add(wt1); expected.add(wt2);
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		ResultSet results = mock(ResultSet.class);
@@ -89,7 +89,7 @@ public class TestWorkloadDAO {
 		when(results.getInt("id")).thenReturn(wt1.getId(),wt2.getId());
 		when(results.getString("url")).thenReturn(wt1.getURL().toString(),wt2.getURL().toString());
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		List<WorkTask> finalReturn = wldao.retrieveWork(expected.size());
 		assertEquals(finalReturn, expected);
 	}
@@ -100,7 +100,7 @@ public class TestWorkloadDAO {
 		int expectedSize = 4;
 		List<WorkTask> expected = new ArrayList<WorkTask>();
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		// Mock methods
@@ -109,7 +109,7 @@ public class TestWorkloadDAO {
 		when(statement.executeUpdate(anyString())).thenReturn(expectedSize);
 		when(statement.executeQuery(anyString())).thenThrow(new SQLException("MOCK SQL ERROR"));
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		List<WorkTask> finalReturn = wldao.retrieveWork(expectedSize);
 		assertEquals(finalReturn, expected);
 	}
@@ -120,7 +120,7 @@ public class TestWorkloadDAO {
 		int expectedSize = 4;
 		List<WorkTask> expected = new ArrayList<WorkTask>();
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		ResultSet results = mock(ResultSet.class);
@@ -130,7 +130,7 @@ public class TestWorkloadDAO {
 		when(statement.executeUpdate(anyString())).thenThrow(new SQLException("MOCK SQL ERROR"));
 		when(statement.executeQuery(anyString())).thenReturn(results);
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		List<WorkTask> finalReturn = wldao.retrieveWork(expectedSize);
 		assertEquals(finalReturn, expected);
 	}
@@ -142,7 +142,7 @@ public class TestWorkloadDAO {
 		WorkTask wt1 = new WorkTask(1, new URL("http://1.com"));
 		expected.add(wt1);
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		ResultSet results = mock(ResultSet.class);
@@ -155,15 +155,15 @@ public class TestWorkloadDAO {
 		when(results.getInt("id")).thenReturn(wt1.getId());
 		when(results.getString("url")).thenReturn(wt1.getURL().toString());
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		List<WorkTask> finalReturn = wldao.retrieveWork(expected.size() + 10); // Ask for 10 more, than received
 		assertEquals(finalReturn, expected);
 	}
 
 	@Test(expected=AssertionError.class)
 	public void testRetrieveWorkNegativeInt() {
-		ConnectionManager connMgr = mock(ConnectionManager.class);
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		wldao.retrieveWork(-1);
 	}
 	
@@ -172,7 +172,7 @@ public class TestWorkloadDAO {
 		int expected = 3;
 		WorkTask wt1 = new WorkTask(1, new URL("http://1.com"));
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		// Mock methods
@@ -180,7 +180,7 @@ public class TestWorkloadDAO {
 		when(conn.createStatement()).thenReturn(statement);
 		when(statement.executeUpdate(anyString())).thenReturn(expected);
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		assertEquals(wldao.checkoutWork(wt1), true);
 		verify(statement).executeUpdate(anyString());
 	}
@@ -189,7 +189,7 @@ public class TestWorkloadDAO {
 	public void testCheckoutWorkSQLException() throws SQLException, MalformedURLException {
 		WorkTask wt1 = new WorkTask(1, new URL("http://1.com"));
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		// Mock methods
@@ -197,7 +197,7 @@ public class TestWorkloadDAO {
 		when(conn.createStatement()).thenReturn(statement);
 		when(statement.executeUpdate(anyString())).thenThrow(new SQLException("MOCK SQL ERROR"));
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		assertEquals(wldao.checkoutWork(wt1), false);
 		verify(statement).executeUpdate(anyString());
 	}
@@ -206,9 +206,9 @@ public class TestWorkloadDAO {
 	public void testCheckoutWorkInvalid() throws SQLException, MalformedURLException {
 		WorkTask wt1 = new WorkTask(1, new URL("http://1.com"));
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		assertFalse(wldao.checkoutWork(wt1));
 	}
 	
@@ -217,7 +217,7 @@ public class TestWorkloadDAO {
 		int expected = 1;
 		URL url1 = new URL("http://1.com");
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		// Mock methods
@@ -225,7 +225,7 @@ public class TestWorkloadDAO {
 		when(conn.createStatement()).thenReturn(statement);
 		when(statement.executeUpdate(anyString(), anyInt())).thenReturn(expected);
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		assertEquals(wldao.submitWork(url1, true), expected);
 	}
 	
@@ -233,7 +233,7 @@ public class TestWorkloadDAO {
 	public void testSubmitWorkSQLException() throws MalformedURLException, SQLException {
 		URL url1 = new URL("http://1.com");
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		// Mock methods
@@ -241,7 +241,7 @@ public class TestWorkloadDAO {
 		when(conn.createStatement()).thenReturn(statement);
 		when(statement.executeUpdate(anyString(), anyInt())).thenThrow(new SQLException("MOCK SQL ERROR"));
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		assertEquals(wldao.submitWork(url1, true), -1);
 	}
 
@@ -249,16 +249,16 @@ public class TestWorkloadDAO {
 	public void testSubmitWorkNull() throws MalformedURLException, SQLException {
 		URL url1 = null;
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		wldao.submitWork(url1, true);		
 	}
 
 	@Test
 	public void testRevertWork() throws SQLException {
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		// Mock methods
@@ -266,14 +266,14 @@ public class TestWorkloadDAO {
 		when(conn.createStatement()).thenReturn(statement);
 		when(statement.executeUpdate(anyString())).thenReturn(1);
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		assertTrue(wldao.revertWork(42)); 
 	}
 	
 	@Test
 	public void testRevertWorkUnknownId() throws SQLException {
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		// Mock methods
@@ -281,7 +281,7 @@ public class TestWorkloadDAO {
 		when(conn.createStatement()).thenReturn(statement);
 		when(statement.executeUpdate(anyString())).thenReturn(0);
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		assertFalse(wldao.revertWork(42)); 
 	}
 	
@@ -289,7 +289,7 @@ public class TestWorkloadDAO {
 	public void testRevertWorkSQLException() throws SQLException {
 		int id = 1;
 		// Mock objects
-		ConnectionManager connMgr = mock(ConnectionManager.class);
+		ConnectionManagerImpl connMgr = mock(ConnectionManagerImpl.class);
 		Connection conn = mock(Connection.class);
 		Statement statement = mock(Statement.class);
 		// Mock methods
@@ -297,7 +297,7 @@ public class TestWorkloadDAO {
 		when(conn.createStatement()).thenReturn(statement);
 		when(statement.executeUpdate(anyString())).thenThrow(new SQLException("MOCK SQL ERROR"));
 		// Run method under inspection
-		WorkloadDAO wldao = new WorkloadDAO(connMgr);
+		WorkloadDaoImpl wldao = new WorkloadDaoImpl(connMgr);
 		assertFalse(wldao.revertWork(id)); 
 	}
 }
